+++ App/World/EntityRegistry.cpp
@@ -5,56 +5,56 @@
 std::vector<glm::mat4> EntityRegistry::allMatrices;
 std::vector<glm::vec3> EntityRegistry::allPositions;
 std::vector<glm::vec3> EntityRegistry::allScales;
-unsigned EntityRegistry::num_entities;
+size_t EntityRegistry::num_entities;
 
 void EntityRegistry::initEntities()
 {
-        unsigned int entities[2];
-	createEntities(2, &entities[0]);
-	setPos(entities[0], glm::vec3( 3.0f, -0.0f, 0.0f));
-        setScale(entities[0], glm::vec3(0.2f, 0.2f, 0.2f));
-        
-        setPos(entities[1], glm::vec3(-3.0f, 0.0f, 0.0f));
-        setScale(entities[1], glm::vec3(1.0f, 0.5f, 1.0f));
-        updateMatrices();
+    size_t entities[2];
+    createEntities(2, &entities[0]);
+    setPos(entities[0], glm::vec3(3.0f, -0.0f, 0.0f));
+    setScale(entities[0], glm::vec3(0.2f, 0.2f, 0.2f));
+
+    setPos(entities[1], glm::vec3(-3.0f, 0.0f, 0.0f));
+    setScale(entities[1], glm::vec3(1.0f, 0.5f, 1.0f));
+    updateMatrices();
 }
 
 void EntityRegistry::updateMatrices()
 {
-        for (unsigned int e = 0; e < num_entities; ++e) {
-                allMatrices[e] = glm::translate(glm::mat4(), allPositions[e]);
-                allMatrices[e] = glm::scale(allMatrices[e], allScales[e]);
-                
-                
-        }
+    for (size_t e = 0; e < num_entities; ++e) {
+	allMatrices[e] = glm::translate(glm::mat4(), allPositions[e]);
+	allMatrices[e] = glm::scale(allMatrices[e], allScales[e]);
+
+
+    }
 }
 
-void EntityRegistry::createEntities(unsigned int pCount, unsigned int * pEntities)
+void EntityRegistry::createEntities(size_t pCount, size_t * pEntities)
 {
-	unsigned int beg = num_entities;
-	for (unsigned int* pEnt = pEntities; num_entities < beg + pCount; ++pEnt) {
-                newEntityID(*pEnt);
-	}
-	allMatrices.resize(num_entities);
-        allPositions.resize(num_entities);
-        allScales.resize(num_entities);
+    size_t beg = num_entities;
+    for (size_t* pEnt = pEntities; num_entities < beg + pCount; ++pEnt) {
+	newEntityID(*pEnt);
+    }
+    allMatrices.resize(num_entities);
+    allPositions.resize(num_entities);
+    allScales.resize(num_entities);
 }
 
-void EntityRegistry::newEntityID(unsigned int& pEntity)
+void EntityRegistry::newEntityID(size_t& pEntity)
 {
-        pEntity = num_entities++;;
+    pEntity = num_entities++;;
 }
 
-void EntityRegistry::setScale(unsigned int pEntityID, glm::vec3 pScale)
+void EntityRegistry::setScale(size_t pEntityID, glm::vec3 pScale)
 {
-        allScales[pEntityID] = pScale;
+    allScales[pEntityID] = pScale;
 }
 
-void EntityRegistry::setPos(unsigned int pEntityID, glm::vec3 pPos)
+void EntityRegistry::setPos(size_t pEntityID, glm::vec3 pPos)
 {
-        allPositions[pEntityID] = pPos;
+    allPositions[pEntityID] = pPos;
 }
-void EntityRegistry::translate(unsigned int pEntityID, glm::vec3 pPos)
+void EntityRegistry::translate(size_t pEntityID, glm::vec3 pPos)
 {
-	allPositions[pEntityID] += pPos;
+    allPositions[pEntityID] += pPos;
 }
