+++ OpenGL/BaseGL/Framebuffer.h
@@ -1,18 +1,18 @@
 #pragma once
 namespace gl {
-	namespace Texture {
-		void initGBuffer();
-		extern unsigned int gBuffer;
-		extern unsigned int gPosTexture;
-		extern unsigned int gNormalTexture;
-		extern unsigned int gAmbientTexture;
-		extern unsigned int gDiffuseTexture;
-		extern unsigned int gSpecularTexture;
-                extern unsigned int gDepthRenderbuffer;
-                void initGUIFBO();
-                extern unsigned int guiFBO;
-                extern unsigned int buttonIndexTexture;
-                extern unsigned int fontColorTexture;
-                extern unsigned int guiDepthRenderbuffer;
-	}
+    namespace Texture {
+	void initGBuffer();
+	extern size_t gBuffer;
+	extern size_t gPosTexture;
+	extern size_t gNormalTexture;
+	extern size_t gAmbientTexture;
+	extern size_t gDiffuseTexture;
+	extern size_t gSpecularTexture;
+	extern size_t gDepthRenderbuffer;
+	void initGUIFBO();
+	extern size_t guiFBO;
+	extern size_t buttonIndexTexture;
+	extern size_t fontColorTexture;
+	extern size_t guiDepthRenderbuffer;
+    }
 }
