+++ OpenGL/BaseGL/Texture.h
@@ -2,43 +2,43 @@
 #include "..\Global\gl.h"
 #include <unordered_map>
 namespace gl {
-	namespace Texture {
+    namespace Texture {
 
-		struct Texture2D {
-			Texture2D(unsigned int pWidth, unsigned int pHeight, GLenum pInternalFormat, GLenum pFormat, GLenum pType);
-			unsigned int ID = 0;
-			unsigned int width = 0;
-			unsigned int height = 0;
-			GLenum internalFormat;
-			GLenum format;
-			GLenum type;
-		};
-		
-		struct TextureBuffer {
-		        TextureBuffer():data(nullptr), width(0), height(0), channels(0) {}
-		        unsigned char* data;
-		        int width;
-		        int height;
-		        int channels;
-		};
-		
-		struct TexData2D{
-			std::vector<unsigned char> data;
-			unsigned int width;
-			unsigned int height;
-		};
-		extern vector<Texture2D> all2DTextures;
-		
-		void loadTextureBuffer(TextureBuffer& pBuffer, std::string pFilename, int pForceChannels = 0);
-		unsigned int createTexture2D(unsigned int pWidth, unsigned int pHeight, GLenum pInternalFormat, GLenum pFormat, GLenum pType, const void* pData);
-		unsigned int createTexture2D(std::string& pFilename);
-		unsigned int createTexture2D(TextureBuffer pBuffer);
-		void setTextureWrapping(unsigned int pTextureIndex, unsigned int pWrapS, unsigned int pWrapT);
-		void setTextureFilter(unsigned int pTextureIndex, unsigned int pMagFilter, unsigned int pMinFilter);
-		
-		void setTextureDirectory(std::string& pDirectory);
-		void resetTextureDirectory();
-		extern std::string TEXTURE_DIR;
-		extern std::unordered_map<std::string, unsigned int> textureLookup;
-	}
+	struct Texture2D {
+	    Texture2D(size_t pWidth, size_t pHeight, GLenum pInternalFormat, GLenum pFormat, GLenum pType);
+	    size_t ID = 0;
+	    size_t width = 0;
+	    size_t height = 0;
+	    GLenum internalFormat;
+	    GLenum format;
+	    GLenum type;
+	};
+
+	struct TextureBuffer {
+	    TextureBuffer() :data(nullptr), width(0), height(0), channels(0) {}
+	    unsigned char* data;
+	    int width;
+	    int height;
+	    int channels;
+	};
+
+	struct TexData2D {
+	    std::vector<unsigned char> data;
+	    size_t width;
+	    size_t height;
+	};
+	extern vector<Texture2D> all2DTextures;
+
+	void loadTextureBuffer(TextureBuffer& pBuffer, std::string pFilename, int pForceChannels = 0);
+	size_t createTexture2D(size_t pWidth, size_t pHeight, GLenum pInternalFormat, GLenum pFormat, GLenum pType, const void* pData);
+	size_t createTexture2D(std::string& pFilename);
+	size_t createTexture2D(TextureBuffer pBuffer);
+	void setTextureWrapping(size_t pTextureIndex, size_t pWrapS, size_t pWrapT);
+	void setTextureFilter(size_t pTextureIndex, size_t pMagFilter, size_t pMinFilter);
+
+	void setTextureDirectory(std::string& pDirectory);
+	void resetTextureDirectory();
+	extern std::string TEXTURE_DIR;
+	extern std::unordered_map<std::string, size_t> textureLookup;
+    }
 }
