+++ OpenGL/GUI/UI/Colored_Quads.h
@@ -2,13 +2,13 @@
 #include <vector>
 #include <string>
 #include <glm/glm.hpp>
-namespace gl{
-        namespace GUI{
-                void colorQuad(unsigned int pQuad, unsigned int pColor);
-                void initColoredQuadVAO();
-                void renderColoredQuads();
-                void initColoredQuadShader();
-                void updateColoredQuads();
-		void setupColoredQuadShader();
-        }
+namespace gl {
+    namespace GUI {
+	void colorQuad(size_t pQuad, size_t pColor);
+	void initColoredQuadVAO();
+	void renderColoredQuads();
+	void initColoredQuadShader();
+	void updateColoredQuads();
+	void setupColoredQuadShader();
+    }
 }
